requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2024.1
psycopg2-binary>=2.9.9
